LayoutTests/http/tests/ssl/applepay/ApplePayCancelEvent.https-expected.txt

Test ApplePaySession's cancel event.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS event is non-null.
PASS typeof event is "object"
PASS event.sessionError is non-null.
PASS typeof event.sessionError is "object"
PASS event.sessionError.code is "unknown"
PASS typeof event.sessionError.code is "string"
PASS Object.keys(event.sessionError.info).length is 0
PASS typeof event.sessionError.info is "object"
PASS cancel event received
PASS successfullyParsed is true

TEST COMPLETE
